refactor(ops): consolidate horizontal scaling instance helpers
